+++ tests/test_appcontrol.py
@@ -32,26 +32,27 @@ def ensure_app_closed(app):
         app.close()
 
 
-def test_native_app_open(app):
+def test_app_open(app):
     assert app.open() is True
-    # assert app.close() is True
-
-
-def test_native_app_is_running(app):
-    # assert app.open() is True
-    # Windows needs some time to report is_running correctly
-    if sys.platform == "win32":
-        time.sleep(1)
-    assert app.is_running() is True
-    # assert app.close() is True
-
-
-def test_native_app_close(app):
-    # assert app.open() is True
     assert app.close() is True
 
 
-def test_native_app_version(app):
+def test_app_is_running(app):
+    assert app.open() is True
+
+    # Windows needs some time to report is_running correctly
+    if sys.platform == "win32":
+        time.sleep(5)
+    assert app.is_running() is True
+    assert app.close() is True
+
+
+def test_app_close(app):
+    assert app.open() is True
+    assert app.close() is True
+
+
+def test_app_version(app):
     assert app.get_version() is not None
 
 
@@ -77,11 +78,11 @@ def test_to_dict(app):
     assert isinstance(app_dict["version"], str)
 
 
-def test_process_info(app, ensure_app_closed):
+def test_process_info(app):
     assert app.open() is True
     # Windows needs some time to report is_running correctly
     if sys.platform == "win32":
-        time.sleep(1)
+        time.sleep(5)
     assert app.is_running() is True
     proc_info = app.process_info()
     assert isinstance(proc_info, dict)